refactor(games): split messages/logging and alliance processing from games/mod.rs
