+++ game/src/games/mod.rs
@@ -16,6 +16,9 @@ use std::collections::HashMap;
 use std::fmt::Display;
 use uuid::Uuid;
 
+pub mod alliances;
+pub mod messages;
+
 /// Stamina restored per phase to a sleeping tribute (PR2c.1, bd-9sjj).
 /// Placeholder pending observability tuning per spec
 /// `2026-05-03-four-phase-day-design.md` §6.4.
@@ -418,145 +421,6 @@ impl Game {
             .filter(|a| !a.is_open())
             .cloned()
             .collect()
-    }
-
-    /// Construct a fallback `MessagePayload` when a caller doesn't supply
-    /// a typed payload. Picks an existing variant suited to the message
-    /// source so the schema-required `payload` field is always present.
-    /// This is a transitional helper used by the legacy log helpers
-    /// pending full migration of every emission site to typed payloads.
-    fn fallback_payload(
-        source: &crate::messages::MessageSource,
-    ) -> crate::messages::MessagePayload {
-        use crate::messages::{AreaEventKind, AreaRef, MessagePayload, MessageSource, TributeRef};
-        match source {
-            MessageSource::Tribute(id) => MessagePayload::SanityBreak {
-                tribute: TributeRef {
-                    identifier: id.clone(),
-                    name: String::new(),
-                },
-            },
-            MessageSource::Area(name) => MessagePayload::AreaEvent {
-                area: AreaRef {
-                    identifier: name.clone(),
-                    name: name.clone(),
-                },
-                kind: AreaEventKind::Other,
-                description: String::new(),
-            },
-            MessageSource::Game(_) => MessagePayload::GameEnded { winner: None },
-        }
-    }
-
-    /// Build and push a `GameMessage` with the supplied typed payload.
-    /// Stamps `(game_day, phase, tick, emit_index)` from the game's
-    /// transient cycle state. The `tick` argument is supplied by the
-    /// caller because some sites (cycle announcements, area events)
-    /// emit at the phase boundary (`tick = 0`) while per-tribute
-    /// action emissions advance the tick counter.
-    pub(crate) fn push_message(
-        &mut self,
-        source: crate::messages::MessageSource,
-        subject: String,
-        content: String,
-        payload: crate::messages::MessagePayload,
-        tick: u32,
-    ) {
-        let game_day = self.day.unwrap_or(0);
-        // Always prefix subject with the game identifier so that the
-        // API's per-game log queries (`WHERE string::starts_with(subject,
-        // $game_id)`) match every emitted message regardless of source
-        // type (Game / Area / Tribute). Without this prefix, area and
-        // tribute messages would be invisible to the day-page and
-        // timeline-summary endpoints.
-        let scoped_subject = if subject.starts_with(&format!("{}:", self.identifier)) {
-            subject
-        } else {
-            format!("{}:{}", self.identifier, subject)
-        };
-        let msg = crate::messages::GameMessage::new(
-            source,
-            game_day,
-            self.current_phase,
-            tick,
-            self.emit_index,
-            scoped_subject,
-            content,
-            payload,
-        );
-        self.messages.push(msg);
-        self.emit_index = self.emit_index.saturating_add(1);
-    }
-
-    /// Push a message into the cycle's transient event buffer.
-    /// The API layer drains and persists this buffer after each cycle.
-    ///
-    /// This legacy helper synthesises a fallback payload (see
-    /// `fallback_payload`) suited to the source. New emission sites
-    /// should construct a typed `MessagePayload` and call
-    /// [`Self::push_message`] directly.
-    pub fn log(
-        &mut self,
-        source: crate::messages::MessageSource,
-        subject: String,
-        content: String,
-    ) {
-        let payload = Self::fallback_payload(&source);
-        let tick = self.tick_counter.boundary();
-        self.push_message(source, subject, content, payload, tick);
-    }
-
-    /// Log a structured game output by rendering its `Display` impl into a `GameMessage`.
-    pub fn log_output<D: std::fmt::Display>(
-        &mut self,
-        source: crate::messages::MessageSource,
-        subject: String,
-        output: D,
-    ) {
-        self.log(source, subject, output.to_string());
-    }
-
-    /// Legacy helper: log a string output and tag with a typed `MessageKind`.
-    /// `kind` is now derived from `MessagePayload::kind()` so the explicit
-    /// `kind` argument is ignored — the variant of the synthesised
-    /// fallback payload determines the kind. New sites should construct
-    /// a typed `MessagePayload` and call [`Self::push_message`] directly.
-    pub fn log_output_kind<D: std::fmt::Display>(
-        &mut self,
-        source: crate::messages::MessageSource,
-        subject: String,
-        output: D,
-        _kind: crate::messages::MessageKind,
-    ) {
-        self.log(source, subject, output.to_string());
-    }
-
-    /// Log a structured [`crate::events::GameEvent`] by rendering its
-    /// `Display` impl into the `GameMessage.content`. The typed
-    /// `MessagePayload` defaults to a source-appropriate fallback;
-    /// callers needing a specific payload variant should instead build
-    /// it themselves and call [`Self::push_message`].
-    pub fn log_event(
-        &mut self,
-        source: crate::messages::MessageSource,
-        subject: String,
-        event: crate::events::GameEvent,
-    ) {
-        self.log(source, subject, event.to_string());
-    }
-
-    /// Log a structured `GameEvent` and tag with `MessageKind`.
-    /// As with [`Self::log_output_kind`], the `kind` argument is now
-    /// derived from the payload and is accepted for backwards
-    /// compatibility only.
-    pub fn log_event_kind(
-        &mut self,
-        source: crate::messages::MessageSource,
-        subject: String,
-        event: crate::events::GameEvent,
-        _kind: crate::messages::MessageKind,
-    ) {
-        self.log(source, subject, event.to_string());
     }
 
     fn check_for_winner(&mut self) -> Result<(), GameError> {
@@ -1152,261 +1016,6 @@ impl Game {
             }
         }
         Ok(())
-    }
-
-    /// Drain the alliance event queue accumulated during the current cycle.
-    /// Called between tribute turns inside `run_tribute_cycle` so cascades
-    /// resolve before the next tribute acts. Per spec §7.5:
-    /// - `BetrayalRecorded`: remove the symmetric pair on the victim's side
-    ///   (betrayer's side was already cleaned at trigger time) and flag the
-    ///   victim for a trust-shock roll on their next turn. The betrayer is
-    ///   never flagged.
-    /// - `DeathRecorded`: roll a sanity-break per direct ally of the deceased
-    ///   (consistent with §7.3a thresholds) and emit a break message on
-    ///   success. After the cascade, unconditionally scrub the deceased's
-    ///   id from every surviving tribute's `allies` list.
-    pub fn process_alliance_events(&mut self, rng: &mut impl Rng) {
-        use crate::tributes::alliances::{AllianceEvent, sanity_break_roll};
-
-        // Collect drained events into a local Vec so we can release the
-        // borrow on `self.alliance_events` before mutating `self.tributes`.
-        let drained: Vec<AllianceEvent> = self.alliance_events.drain(..).collect();
-
-        for ev in drained {
-            match ev {
-                AllianceEvent::BetrayalRecorded { betrayer, victim } => {
-                    // Snapshot names before the mutable borrow on `victim` so
-                    // we can emit the message after victim mutation completes.
-                    let betrayer_info = self
-                        .tributes
-                        .iter()
-                        .find(|t| t.id == betrayer)
-                        .map(|t| (t.identifier.clone(), t.name.clone()));
-                    let victim_info = self
-                        .tributes
-                        .iter()
-                        .find(|t| t.id == victim)
-                        .map(|t| (t.identifier.clone(), t.name.clone()));
-                    if let Some(v) = self.tributes.iter_mut().find(|t| t.id == victim) {
-                        v.allies.retain(|x| *x != betrayer);
-                        v.pending_trust_shock = true;
-                    }
-                    // Spec §7.5: betrayer is never enqueued for trust-shock.
-                    if let (Some((b_id, b_name)), Some((v_id, v_name))) =
-                        (betrayer_info, victim_info)
-                    {
-                        let event = crate::events::GameEvent::BetrayalTriggered {
-                            betrayer_id: betrayer,
-                            betrayer_name: b_name.clone(),
-                            victim_id: victim,
-                            victim_name: v_name.clone(),
-                        };
-                        let payload = crate::messages::MessagePayload::BetrayalTriggered {
-                            betrayer: crate::messages::TributeRef {
-                                identifier: b_id,
-                                name: b_name,
-                            },
-                            victim: crate::messages::TributeRef {
-                                identifier: v_id.clone(),
-                                name: v_name.clone(),
-                            },
-                        };
-                        let tick = self.tick_counter.next();
-                        self.push_message(
-                            crate::messages::MessageSource::Tribute(v_id),
-                            v_name,
-                            event.to_string(),
-                            payload,
-                            tick,
-                        );
-                    }
-                }
-                AllianceEvent::DeathRecorded {
-                    deceased,
-                    killer: _,
-                } => {
-                    // Snapshot the deceased's allies and identifying refs
-                    // before mutation so we can roll the cascade per direct
-                    // ally and emit typed `TrustShockBreak` payloads.
-                    let (allies_of_deceased, deceased_ref): (Vec<Uuid>, _) = self
-                        .tributes
-                        .iter()
-                        .find(|t| t.id == deceased)
-                        .map(|d| {
-                            (
-                                d.allies.clone(),
-                                crate::messages::TributeRef {
-                                    identifier: d.identifier.clone(),
-                                    name: d.name.clone(),
-                                },
-                            )
-                        })
-                        .unwrap_or_else(|| {
-                            (
-                                Vec::new(),
-                                crate::messages::TributeRef {
-                                    identifier: deceased.to_string(),
-                                    name: String::new(),
-                                },
-                            )
-                        });
-
-                    for ally_id in allies_of_deceased {
-                        if let Some(ally) = self.tributes.iter_mut().find(|t| t.id == ally_id) {
-                            // `extreme_low_sanity` is the §7.3a low-limit
-                            // mapping (see PersonalityThresholds doc).
-                            let limit = ally.brain.thresholds.extreme_low_sanity;
-                            let sanity = ally.attributes.sanity;
-                            if sanity_break_roll(sanity, limit, rng) {
-                                ally.allies.retain(|x| *x != deceased);
-                                let aid = ally.identifier.clone();
-                                let aname = ally.name.clone();
-                                let ally_uuid = ally.id;
-                                let event = crate::events::GameEvent::TrustShockBreak {
-                                    tribute_id: ally_uuid,
-                                    tribute_name: aname.clone(),
-                                };
-                                let payload = crate::messages::MessagePayload::TrustShockBreak {
-                                    tribute: crate::messages::TributeRef {
-                                        identifier: aid.clone(),
-                                        name: aname.clone(),
-                                    },
-                                    partner: deceased_ref.clone(),
-                                };
-                                let tick = self.tick_counter.next();
-                                self.push_message(
-                                    crate::messages::MessageSource::Tribute(aid),
-                                    aname,
-                                    event.to_string(),
-                                    payload,
-                                    tick,
-                                );
-                            }
-                        }
-                    }
-
-                    // Unconditional cleanup: ensure the deceased's id is
-                    // removed from every surviving tribute's allies list,
-                    // even if their cascade roll failed.
-                    for t in self.tributes.iter_mut() {
-                        t.allies.retain(|x| *x != deceased);
-                    }
-                }
-                AllianceEvent::FormationRecorded {
-                    proposer,
-                    target,
-                    factor,
-                } => {
-                    let proposer_info = self
-                        .tributes
-                        .iter()
-                        .find(|t| t.id == proposer)
-                        .map(|t| (t.identifier.clone(), t.name.clone()));
-                    let target_info = self
-                        .tributes
-                        .iter()
-                        .find(|t| t.id == target)
-                        .map(|t| (t.identifier.clone(), t.name.clone()));
-                    let mut idx_p: Option<usize> = None;
-                    let mut idx_t: Option<usize> = None;
-                    for (i, t) in self.tributes.iter().enumerate() {
-                        if t.id == proposer {
-                            idx_p = Some(i);
-                        }
-                        if t.id == target {
-                            idx_t = Some(i);
-                        }
-                    }
-                    let (Some(ip), Some(it)) = (idx_p, idx_t) else {
-                        continue;
-                    };
-                    if self.tributes[ip].allies.len() >= crate::tributes::alliances::MAX_ALLIES
-                        || self.tributes[it].allies.len() >= crate::tributes::alliances::MAX_ALLIES
-                    {
-                        continue;
-                    }
-                    if !self.tributes[ip].allies.contains(&target) {
-                        self.tributes[ip].allies.push(target);
-                    }
-                    if !self.tributes[it].allies.contains(&proposer) {
-                        self.tributes[it].allies.push(proposer);
-                    }
-                    if let (Some((p_id, p_name)), Some((t_id, t_name))) =
-                        (proposer_info, target_info)
-                    {
-                        let event = crate::events::GameEvent::AllianceFormed {
-                            tribute_a_id: proposer,
-                            tribute_a_name: p_name.clone(),
-                            tribute_b_id: target,
-                            tribute_b_name: t_name.clone(),
-                            factor: factor.clone(),
-                        };
-                        let payload = crate::messages::MessagePayload::AllianceFormed {
-                            members: vec![
-                                crate::messages::TributeRef {
-                                    identifier: p_id.clone(),
-                                    name: p_name.clone(),
-                                },
-                                crate::messages::TributeRef {
-                                    identifier: t_id,
-                                    name: t_name,
-                                },
-                            ],
-                        };
-                        let tick = self.tick_counter.next();
-                        self.push_message(
-                            crate::messages::MessageSource::Tribute(p_id),
-                            p_name,
-                            event.to_string(),
-                            payload,
-                            tick,
-                        );
-                    }
-                }
-                AllianceEvent::AllianceSummons { summoner, target } => {
-                    // Spec §6.4 PR2c.2 (bd-1zju). When the target is asleep,
-                    // an ally's summons interrupts the rest. Currently no
-                    // production code emits this event; the handler is in
-                    // place so future PRs (or test scaffolding) can wake
-                    // sleeping allies through the standard alliance pipeline.
-                    let summoner_ref = self.tributes.iter().find(|t| t.id == summoner).map(|t| {
-                        crate::messages::TributeRef {
-                            identifier: t.identifier.clone(),
-                            name: t.name.clone(),
-                        }
-                    });
-                    let Some(s_ref) = summoner_ref else { continue };
-                    let phase = self.current_phase;
-                    let mut wake_events: Vec<crate::messages::TaggedEvent> = Vec::new();
-                    let woke_info =
-                        if let Some(t) = self.tributes.iter_mut().find(|t| t.id == target) {
-                            if t.wake_interrupted(
-                                shared::messages::InterruptionKind::AllianceSummons { ally: s_ref },
-                                phase,
-                                &mut wake_events,
-                            ) {
-                                Some((t.identifier.clone(), t.name.clone()))
-                            } else {
-                                None
-                            }
-                        } else {
-                            None
-                        };
-                    if let Some((t_id, t_name)) = woke_info {
-                        for ev in wake_events.drain(..) {
-                            let tick = self.tick_counter.next();
-                            self.push_message(
-                                crate::messages::MessageSource::Tribute(t_id.clone()),
-                                t_name.clone(),
-                                ev.content,
-                                ev.payload,
-                                tick,
-                            );
-                        }
-                    }
-                }
-            }
-        }
     }
 
     /// Spawn one sponsor per archetype using the shared catalog.
